backend/requirements.txt
fastapi
uvicorn
openai
python-dotenv
langchain
langchain[llms]
langchain-core
